+++ lib/iord/resource_url.rb
@@ -5,10 +5,17 @@ module Iord
     extend ActiveSupport::Concern
 
     included do
+      helper_method :has_collection?
+
+      helper_method :form_resource_url
       helper_method :resource_url
       helper_method :collection_url
       helper_method :new_resource_url
       helper_method :edit_resource_url
+    end
+
+    def has_collection?
+      @has_collection ||= self.respond_to? collection_url_method
     end
 
     def resource_url_method
@@ -16,11 +23,21 @@ module Iord
     end
 
     def collection_url_method
-      @collection_url_method ||= (action_path + '_url').to_sym
+      @collection_url_method ||= (action_path.pluralize + '_url').to_sym
+    end
+
+    def form_resource_url
+      if @resource.persisted? or not has_collection?
+        resource_url
+      else
+        collection_url
+      end
     end
 
     def resource_url(resource = nil)
-      self.public_send resource_url_method.to_sym, (resource || @resource)
+      resource ||= @resource
+      resource = nil unless has_collection?
+      self.public_send resource_url_method.to_sym, resource
     end
 
     def collection_url
@@ -32,7 +49,9 @@ module Iord
     end
 
     def edit_resource_url(resource = nil)
-      self.public_send "edit_#{resource_url_method}".to_sym, (resource || @resource)
+      resource ||= @resource
+      resource = nil unless has_collection?
+      self.public_send "edit_#{resource_url_method}".to_sym, resource
     end
   end
 end