+++ test/dummy/app/controllers/admin/administrators_controller.rb
@@ -1,0 +1,8 @@
+require 'iord/controller'
+
+module Admin
+  class AdministratorsController < ApplicationController
+    include Iord::Controller
+  end
+end
+