+++ test/dummy/app/controllers/administrators_controller.rb
@@ -1,0 +1,6 @@
+require 'iord/controller'
+
+class AdministratorsController < ApplicationController
+  include Iord::Controller
+end
+